Moving fetch tests out
